#121 Extract Proxy Credential Facotry from Intermediate Class
