+++ src/main/java/io/securecodebox/persistence/defectdojo/http/Foo.java
@@ -8,19 +8,10 @@ import io.securecodebox.persistence.defectdojo.config.Config;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.http.HttpHost;
-import org.apache.http.auth.AuthScope;
-import org.apache.http.auth.Credentials;
-import org.apache.http.auth.UsernamePasswordCredentials;
-import org.apache.http.client.CredentialsProvider;
-import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.impl.client.ProxyAuthenticationStrategy;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.web.client.RestTemplate;
-
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 
 /**
  * Placeholder to move duplicated code, will be named better later
@@ -36,16 +27,13 @@ public final class Foo {
     this.proxyConfig = proxyConfig;
   }
 
-
-
-
   public RestTemplate createRestTemplate() {
     if (proxyConfig.isComplete()) {
       // Configuring Proxy Authentication explicitly as it isn't done by default for spring rest templates :(
       final var builder = HttpClientBuilder.create()
         .useSystemProperties()
         .setProxy(createHttpHost())
-        .setDefaultCredentialsProvider(createCredentialsProvider())
+        .setDefaultCredentialsProvider(new ProxyCredentialFactory(proxyConfig).createCredentialsProvider())
         .setProxyAuthenticationStrategy(new ProxyAuthenticationStrategy());
 
       final var factory = new HttpComponentsClientHttpRequestFactory();
@@ -55,20 +43,6 @@ public final class Foo {
     }
 
     return new RestTemplate();
-  }
-
-  CredentialsProvider createCredentialsProvider() {
-    final var provider = new BasicCredentialsProvider();
-    provider.setCredentials(createAuthScope(), createCredentials());
-    return provider;
-  }
-
-  AuthScope createAuthScope() {
-    return new AuthScope(proxyConfig.getHost(), proxyConfig.getPort());
-  }
-
-  Credentials createCredentials() {
-    return new UsernamePasswordCredentials(proxyConfig.getUser(), proxyConfig.getPassword());
   }
 
   HttpHost createHttpHost() {
